fixed favorites contacts select and deselect

+++ shotvibe/Delegates/SVAddressBook.h
@@ -23,7 +23,7 @@ typedef void(^AddressBookSearchCompletionBlock)();
 
 - (id)initWithBlock:(AddressBookPermissionsBlock)completionBlock;
 - (void)filterByKeyword:(NSString*)keyword completionBlock:(AddressBookSearchCompletionBlock)completionBlock;
-- (int)idOfRecord:(ABRecordRef)record;
+- (int64_t)idOfRecord:(ABRecordRef)record;
 - (ABRecordRef)recordOfRecordId:(ABRecordID)recordId;
 
 @end
